+++ src/main/java/com/gmail/val59000mc/configuration/MainConfig.java
@@ -142,6 +142,7 @@ public class MainConfig extends YamlFile {
 	public static final BasicOption<Boolean> ENABLE_TEAM_NAMES = new BasicOption<>("enable-team-names",true);
 	public static final BasicOption<Integer> TIME_BEFORE_RESTART_AFTER_END = new BasicOption<>("time-before-restart-after-end",30);
 	public static final BasicOption<Boolean> CAN_JOIN_AS_SPECTATOR = new BasicOption<>("can-join-as-spectator",false);
+	public static final BasicOption<Boolean> COUNT_OFFLINE_TEAMS_AS_DEAD = new BasicOption<>("count-offline-teams-as-dead",true);
 	public static final BasicOption<Boolean> END_GAME_WHEN_ALL_PLAYERS_HAVE_LEFT = new BasicOption<>("countdown-ending-game-when-all-players-have-left",true);
 	public static final BasicOption<Integer> DELAY_BETWEEN_TEAM_TELEPORTS = new BasicOption<>("delay-between-team-teleports", 10);
 	public static final BasicOption<Boolean> DEBUG = new BasicOption<>("debug",false);